TST add fft tests, seperate numpy backend tests from scattering tests
